requirements.txt
# world generator
portkey-ai
openai
python-dotenv
langchain-core
langchain-openai
pydantic
flask
pillow